packages/neo-one-client/src/utils.ts
import { common, ScriptBuilder, ScriptBuilderParam } from '@neo-one/client-core';
import { Hash160String } from './types';

export const getInvokeMethodInvocationScript = ({
  method,
  params,
}: {
  readonly method: string;
  readonly params: ReadonlyArray<ScriptBuilderParam | undefined>;
}): Buffer => {
  const sb = new ScriptBuilder();
  sb.emitAppCallInvocation(method, ...params);

  return sb.build();
};

export const getInvokeMethodScript = ({
  hash,
  method,
  params,
}: {
  readonly hash: Hash160String;
  readonly method: string;
  readonly params: ReadonlyArray<ScriptBuilderParam | undefined>;
}): Buffer => {
  const sb = new ScriptBuilder();
  sb.emitAppCall(common.stringToUInt160(hash), method, ...params);

  return sb.build();
};
